+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Implement i18n with support to english and portuguese [#2](https://github.com/BIEMAX/financial-manager-app/issues/2)
 - Create a default structure for i18n messages [#4](https://github.com/BIEMAX/quote-manager/issues/4)
 - Fix menu 'Task todo' not accessible [#7](https://github.com/BIEMAX/quote-manager/issues/7)
+- Add new sub routes for finances [#8](https://github.com/BIEMAX/quote-manager/issues/8)